aiohttp-container/pyproject.toml
[project]
name = "customer"
version = "0.1.0"
description = "MCP server to provide customer interactions"
authors = [
    {name = "Ben Greene",email = "BenJGreene@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
[tool.poetry.dependencies]

click =  "^8"
pydantic-yaml = {extras = ["pyyaml"], version = "^1.4"}
prometheus-async = "^25.1"
pydantic-file-secrets = "^0.4"
pydantic-settings = {extras = ["yaml"], version = "^2.6.1"}
aiohttp-mcp = "^0.3.0"


[tool.poetry.group.dev.dependencies]
aiohttp-devtools = "^1"
pytest = "^8"
Sphinx = "^8"
sphinx-autodoc-typehints = "^3.2.0"
pytest-sphinx = "^0.6"
flake8 = "^7.1.1"
flake8-black = "^0.3.6"
pylint = "^3.3.1"
sphinx-click = "^6.0.0"
pytest-click = "^1.1.0"
pytest-cov = "^6.1"
pytest-watch = "^4.2.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
customer = "customer.cli:cli"
